metadata: implement journal tombstones
Thombstones are used to represent deleted files in the journal. A mode
of zero (no permission bits set) is used to signify deleted files. This
is not expected to be a likely file mode, and in the event that a real
file in the repository exists with such a mode, then there is nothing
of interest to store in the overlay. The expectation is that the file
metadata can be regenerated from the underlying filesystem in these
rare cases.

Signed-off-by: Simeon Miteff <simeon.miteff@gmail.com>
Signed-off-by: Edwin Peer <espeer@gmail.com>